Core components
@jdelara @patrickneubauer @manuelWimmer

Hi all,

I am updating the tool with the last changes (custom queries and
predicated based queries) introduced in the main contribution section
of the paper. This update belongs to the changes that affects to the
metamodel. I will try to have a new update site for the tool as soon as
possible.

Cheers from Madrid,

Ángel