Several issues fixed
@patrickneubauer @jdelara @manuelWimmer

Hi all,

Patrick and I are finishing with the work. I have updated the tool with
the issues Patrick detected with his case study and prepare a new query
for him. Actually, I have updated the paper too in order to recover the
table with the LOCs for the evaluation and added the integration with
Exeed.

Hi have updated the website

http://angel539.github.io/extremo/

with the new information provided in the paper.

All the best,

Ángel